requirements_flask_legacy.txt
line-bot-sdk==3.8.0
flask
flask-restful
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-cors
flask-marshmallow
marshmallow-sqlalchemy
python-dotenv
openai
google-genai
requests
markdown
deepgram-sdk>=4.0.0
faster-whisper
psycopg2-binary
bcrypt
celery
redis
uuid
werkzeug